Refactors responses dialogs initial work
