+++ components/form-builder/app/responses/Dialogs/line-item-entries/LineItemEntries.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef } from "react";
 import { LineItems } from "./LineItems";
 import { scrollToBottom } from "@lib/clientHelpers";
 import { useTranslation } from "react-i18next";
-import { DialogStates } from "../../responses/DownloadTableDialog";
+import { DialogStates } from "../ConfirmDialog";
 
 // TODO: handle duplicate entries?
 // TODO: should "backspace" on an empty input set the next entry into "edit mode"?
