+++ components/form-builder/app/shared/index.ts
@@ -13,7 +13,6 @@ export { TagInput } from "./tag-input";
 export { LockedBadge } from "./LockedBadge";
 export { ExclamationText } from "./ExclamationText";
 export { InfoDetails } from "./InfoDetails";
-export { LineItem, LineItems, LineItemEntries } from "./line-item-entries";
 export { ResponseDeliveryHelpButton } from "./ResponseDeliveryHelpDialog";
 export { toast } from "./Toast";
 export { EmailResponseSettings } from "./EmailResponseSettings";
